slip invoice

+++ src/main/java/com/bprasojo/ekspedisi/dao/TagihanCustomerDAO.java
@@ -13,7 +13,9 @@ import com.bprasojo.ekspedisi.model.TagihanCustomer;
 import java.sql.*;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -35,7 +37,7 @@ public class TagihanCustomerDAO {
         // Query untuk mencari nomor bukti tertinggi yang memiliki awalan "BON-" + tahun + bulan yang sama
         String sql = "SELECT MAX(SUBSTRING(no_invoice, 11)) AS last_number " +
                      "FROM tagihan_customer " +
-                     "WHERE no_invoice LIKE 'INVs-" + yearMonth + "%'";
+                     "WHERE no_invoice LIKE 'INV-" + yearMonth + "%'";
 
         try (Statement stmt = conn.createStatement(); 
             ResultSet rs = stmt.executeQuery(sql)) {
@@ -68,6 +70,10 @@ public class TagihanCustomerDAO {
     public void save(TagihanCustomer tagihan) throws SQLException {
         String sql;
         
+        if (tagihan.getPelunasan() > 0){
+            throw new SQLException("Invoice tidak bisa diubah karena sudah ada pembayaran");
+        }
+        
         if (tagihan.getNilaiPekerjaan() <= tagihan.getPelunasan()){
             tagihan.setStatusLunas("lunas");
         } else {
@@ -77,10 +83,10 @@ public class TagihanCustomerDAO {
         boolean isInsert = tagihan.getId() == 0; 
         if (isInsert) {
             tagihan.setNoInvoice(generateNoInvoice(tagihan.getTanggal()));
-            sql = "INSERT INTO tagihan_customer (customer_id, no_invoice, tanggal, pekerjaan, nilai_pekerjaan, ppn_persen, ppn, total, terbilang, pelunasan, status_lunas) " +
-                  "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
+            sql = "INSERT INTO tagihan_customer (customer_id, no_invoice, tanggal, pekerjaan, nilai_pekerjaan, ppn_persen, ppn, total, terbilang, pelunasan, status_lunas, keterangan) " +
+                  "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?,?)";
         } else {
-            sql = "UPDATE tagihan_customer SET customer_id=?, no_invoice=?, tanggal=?, pekerjaan=?, nilai_pekerjaan=?, ppn_persen=?, ppn=?, total=?, terbilang=?, pelunasan=?, status_lunas=? WHERE id=?";
+            sql = "UPDATE tagihan_customer SET customer_id=?, no_invoice=?, tanggal=?, pekerjaan=?, nilai_pekerjaan=?, ppn_persen=?, ppn=?, total=?, terbilang=?, pelunasan=?, status_lunas=?, keterangan=? WHERE id=?";
         }
 
         try (PreparedStatement stmt = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
@@ -97,9 +103,10 @@ public class TagihanCustomerDAO {
             stmt.setString(9, tagihan.getTerbilang());
             stmt.setInt(10, tagihan.getPelunasan());
             stmt.setString(11, tagihan.getStatusLunas());
+            stmt.setString(12, tagihan.getKeterangan());
 
             if (tagihan.getId() != 0) {
-                stmt.setInt(12, tagihan.getId());
+                stmt.setInt(13, tagihan.getId());
             }
 
             stmt.executeUpdate();
@@ -115,6 +122,15 @@ public class TagihanCustomerDAO {
 
     // Hapus data
     public void delete(int id) throws SQLException {
+        TagihanCustomer invoice = getById(id);
+        if (invoice == null){
+            throw new SQLException("Data invoice tidak ditemukan");
+        }
+        
+        if (invoice.getPelunasan() > 0){
+            throw new SQLException("Data invoice tidak bisa dihapus karena sudah ada pembayaran");
+        }
+        
         String sql = "DELETE FROM tagihan_customer WHERE id=?";
         try (PreparedStatement stmt = conn.prepareStatement(sql)) {
             stmt.setInt(1, id);
@@ -123,7 +139,7 @@ public class TagihanCustomerDAO {
     }
 
     // Dapatkan data berdasarkan ID atau No Invoice
-    public TagihanCustomer getByField(String field, Object value) throws SQLException {
+    private TagihanCustomer getByField(String field, Object value) throws SQLException {
         String sql = "SELECT * FROM tagihan_customer WHERE " + field + " = ?";
         try (PreparedStatement stmt = conn.prepareStatement(sql)) {
             if (value instanceof Integer) {
@@ -146,7 +162,8 @@ public class TagihanCustomerDAO {
                         rs.getInt("total"),
                         rs.getString("terbilang"),
                         rs.getInt("pelunasan"),
-                        rs.getString("status_lunas")
+                        rs.getString("status_lunas"),
+                        rs.getString("keterangan")
                     );
                 }
             }
@@ -162,16 +179,16 @@ public class TagihanCustomerDAO {
         return getByField("no_invoice", noInvoice);
     }
     
-    public List<TagihanCustomer> getTagihanCustomerByPage(Integer page, java.util.Date tglAwal, java.util.Date tglAkhir, String filter) throws SQLException {
-        List<TagihanCustomer> tagihanList = new ArrayList<>();
+    public List<Map<String, Object>> getTagihanCustomerByPage(Integer page, java.util.Date tglAwal, java.util.Date tglAkhir, String filter) throws SQLException {
+        List<Map<String, Object>> resultList = new ArrayList<>();
 
         // Query dasar
-        String sql = "SELECT a.*, a.nama as nama_cusomer FROM tagihan_customer a " 
+        String sql = "SELECT a.*, b.nama as nama_customer, 'Lia' as pc FROM tagihan_customer a " 
                      + " inner join stake_holder b on a.customer_id = b.id "
                      + " WHERE a.tanggal BETWEEN ? AND ? "; 
 
         if (filter != null && !filter.trim().isEmpty()) {
-            sql += " WHERE (no_invoice LIKE ? OR pekerjaan LIKE ? or keterangan like ? or b.nama like ?)";
+            sql += " and (no_invoice LIKE ? OR pekerjaan LIKE ? or keterangan like ? or b.nama like ? or a.keterangan like ?)";
         }
 
         sql += " LIMIT ? OFFSET ?";
@@ -184,7 +201,7 @@ public class TagihanCustomerDAO {
 
             // Set filter parameters if present
             if (filter != null && !filter.trim().isEmpty()) {
-                for (int i = 0; i < 4; i++) { // Filter untuk 5 kolom
+                for (int i = 0; i < 5; i++) { // Filter untuk 5 kolom
                     stmt.setString(paramIndex++, "%" + filter + "%");
                 }
             }
@@ -194,28 +211,26 @@ public class TagihanCustomerDAO {
             stmt.setInt(paramIndex++, pageSize); // Parameter untuk LIMIT
             stmt.setInt(paramIndex, (page - 1) * pageSize); // Parameter untuk OFFSET
 
+            // Eksekusi query
             try (ResultSet rs = stmt.executeQuery()) {
+                // Mendapatkan metadata kolom
+                ResultSetMetaData metaData = rs.getMetaData();
+                int columnCount = metaData.getColumnCount();
+
+                // Proses setiap row hasil query
                 while (rs.next()) {
-                    TagihanCustomer tagihan = new TagihanCustomer(
-                        rs.getInt("id"),
-                        rs.getInt("customer_id"),
-                        rs.getString("no_invoice"),
-                        rs.getDate("tanggal"),
-                        rs.getString("pekerjaan"),
-                        rs.getInt("nilai_pekerjaan"),
-                        rs.getInt("ppn_persen"),
-                        rs.getInt("ppn"),
-                        rs.getInt("total"),
-                        rs.getString("terbilang"),
-                        rs.getInt("pelunasan"),
-                        rs.getString("status_lunas")
-                    );
-                    tagihanList.add(tagihan);
+                    Map<String, Object> rowMap = new LinkedHashMap<>();
+                    for (int i = 1; i <= columnCount; i++) {
+                        String columnName = metaData.getColumnLabel(i); // Nama kolom
+                        Object columnValue = rs.getObject(i); // Nilai kolom
+                        rowMap.put(columnName, columnValue); // Menyimpan dalam Map
+                    }
+                    resultList.add(rowMap); // Menambahkan baris ke dalam list
                 }
             }
         }
 
-        return tagihanList;
+        return resultList;
     }
 
 }
